+++ examples/use_rust/rp2040_dfu_split/src/central.rs
@@ -109,6 +109,7 @@ async fn main(_spawner: Spawner) {
         manufacturer: "Haobo",
         product_name: "RMK Keyboard RP2040 embassy-boot split use_rust example",
         serial_number: "vial:f64c2b3c:000001",
+        device_release: 0x0010,
     };
 
     let vial_config = VialConfig::new(VIAL_KEYBOARD_ID, VIAL_KEYBOARD_DEF, &[(0, 0), (1, 1)]);