+++ examples/use_rust/rp2040_embassy_boot_split/src/peripheral.rs
@@ -72,6 +72,7 @@ async fn main(_spawner: Spawner) {
         manufacturer: "Haobo",
         product_name: "RMK Keyboard Peripheral",
         serial_number: "vial:f64c2b3c:000001",
+        device_release: 0x0010,
     };
 
     let mut dfu_led = DfuLedProcessor::new(Output::new(p.PIN_25, Level::Low), false);